+++ run.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 docker rm -f nonotion-service mongo-db
-docker image rm bd2-nonotion-service
+if docker image inspect bd2-nonotion-service > /dev/null 2>&1; then
+    docker image rm bd2-nonotion-service
+fi
 call .\mvnw.cmd clean package -Dmaven.test.skip=false
 docker compose up -d
